# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: TapasHub Business Operating System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [inventory]
      summary: List products
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: search
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated products
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
    post:
      operationId: createProduct
      tags: [inventory]
      summary: Create a product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{productId}:
    get:
      operationId: getProduct
      tags: [inventory]
      summary: Get a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    patch:
      operationId: updateProduct
      tags: [inventory]
      summary: Update a product
      parameters:
        - name: productId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /inventory/low-stock:
    get:
      operationId: getLowStockItems
      tags: [inventory]
      summary: Products below reorder threshold
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Low stock items
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
...
  /finance/transactions:
    get:
      operationId: listTransactions
      tags: [finance]
      summary: List financial transactions
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: type
          in: query
          schema:
            type: ["string", "null"]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Paginated transactions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
    post:
      operationId: createTransaction
      tags: [finance]
      summary: Record a transaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transaction"
  /finance/pnl-summary:
    get:
      operationId: getPnlSummary
      tags: [finance]
      summary: Profit & Loss summary
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
        - name: period
          in: query
          schema:
            type: string
            enum: [month, quarter, year]
            default: month
      responses:
        "200":
          description: P&L summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PnlSummary"
  /finance/cash-flow:
    get:
      operationId: getCashFlow
      tags: [finance]
      summary: Cash flow chart data (last 6 months)
      parameters:
        - name: companyId
          in: query
          schema:
            type: ["integer", "null"]
      responses:
        "200":
          description: Cash flow data
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CashFlowPoint"

  # ─── Employees ───────────────────────────────────────────────────
...
  /notifications:
    get:
      operationId: listNotifications
      tags: [notifications]
      summary: List notifications
      parameters:
        - name: unreadOnly
          in: query
          schema:
            type: boolean
            default: false
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Notifications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Notification"
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Notification"
  /notifications/mark-all-read:
    patch:
      operationId: markAllNotificationsRead
      tags: [notifications]
      summary: Mark all notifications as read
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAllReadInput"
      responses:
        "200":
          description: All marked as read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BulkActionResult"

  # ─── Approvals ───────────────────────────────────────────────────
...
components:
  schemas:
...
    Product:
      type: object
      required: [id, companyId, companyName, name, sku, category, price, costPrice, stockQuantity, reorderLevel, status, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        name:
          type: string
        sku:
          type: string
        barcode:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        price:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        warehouseLocation:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [active, discontinued, draft]
        createdAt:
          type: string
    ProductInput:
      type: object
      required: [companyId, name, sku, category, price, costPrice, stockQuantity, reorderLevel]
      properties:
        companyId:
          type: integer
        name:
          type: string
        sku:
          type: string
        barcode:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        warehouseLocation:
          type: string
    ProductUpdate:
      type: object
      properties:
        name:
          type: string
        category:
          type: string
        description:
          type: string
        price:
          type: number
        costPrice:
          type: number
        stockQuantity:
          type: integer
        reorderLevel:
          type: integer
        warehouseLocation:
          type: string
        status:
          type: string
          enum: [active, discontinued, draft]
    ProductList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
...
    Transaction:
      type: object
      required: [id, companyId, companyName, type, category, amount, description, status, date, createdAt]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        companyName:
          type: string
        type:
          type: string
          enum: [income, expense, transfer]
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        referenceNumber:
          type: ["string", "null"]
        paymentMethod:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, completed, failed, cancelled]
        date:
          type: string
        createdAt:
          type: string
    TransactionInput:
      type: object
      required: [companyId, type, category, amount, description, date]
      properties:
        companyId:
          type: integer
        type:
          type: string
          enum: [income, expense, transfer]
        category:
          type: string
        amount:
          type: number
        description:
          type: string
        referenceNumber:
          type: string
        paymentMethod:
          type: string
        date:
          type: string
    TransactionList:
      type: object
      required: [items, total, page, limit]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Transaction"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    PnlSummary:
      type: object
      required: [revenue, cogs, grossProfit, grossMargin, operatingExpenses, operatingProfit, netProfit, netMargin, period]
      properties:
        revenue:
          type: number
        cogs:
          type: number
        grossProfit:
          type: number
        grossMargin:
          type: number
        operatingExpenses:
          type: number
        operatingProfit:
          type: number
        netProfit:
          type: number
        netMargin:
          type: number
        period:
          type: string
        revenueGrowth:
          type: number
        profitGrowth:
          type: number
    CashFlowPoint:
      type: object
      required: [label, inflow, outflow, net]
      properties:
        label:
          type: string
        inflow:
          type: number
        outflow:
          type: number
        net:
          type: number

    # ─── Employees ─────────────────────────────────────────────────
...
    Notification:
      type: object
      required: [id, type, title, message, isRead, createdAt]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [order, payment, inventory, gst, compliance, hr, marketing, approval, system]
        title:
          type: string
        message:
          type: string
        isRead:
          type: boolean
        companyName:
          type: ["string", "null"]
        actionUrl:
          type: ["string", "null"]
        severity:
          type: string
          enum: [info, warning, error, success]
        createdAt:
          type: string
    MarkAllReadInput:
      type: object
      properties:
        companyId:
          type: ["integer", "null"]
    BulkActionResult:
      type: object
      required: [affected]
      properties:
        affected:
          type: integer

    # ─── Approvals ─────────────────────────────────────────────────
...