# lib/api-zod/src/generated/types/productInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * TapasHub Business Operating System API
 * OpenAPI spec version: 0.1.0
 */

export interface ProductInput {
  companyId: number;
  name: string;
  sku: string;
  barcode?: string;
  category: string;
  description?: string;
  price: number;
  costPrice: number;
  stockQuantity: number;
  reorderLevel: number;
  warehouseLocation?: string;
}